Fix formatting and linting.
In first place, let me apologize. I needed to fix the Finnish formatting
but linting got into my editor's way so this is the complete list of
changes:

- Fix formatting for a few languages. Some languages include a dot in
the short format and in the data we get that dot surrounded by single
quotes. This is the main reason for this PR.
- Make the project pass its linting rules.
- Make CI check linting.
- Fix path in tslint.json to properly exclude test/locale-data.ts
- Add test/locale-data.ts to .prettierignore